Kilimo360_ready/requirements.txt
kivy==2.2.0
kivymd==1.1.1
kivy_garden.mapview
plyer
